+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Both colorscad and 3mfmerge changes are included here. Unless explicitly mention
 - Now it really works on OSX, 0.3.1 actually didn't because of using 'sed -u'
 - A few more sanity checks:
   - Check if 'openscad --info' reports 3MF support
+  - A non-empty .csg is produced during the first step
 
 ## [0.4.0] - 2021-07-04